requirements.txt
# requirements.txt for Easy Import
bcrypt == 4.1.2
blinker == 1.9.0
CacheControl == 0.14.4
cachetools == 6.2.2
certifi == 2025.11.12
cffi == 2.0.0
charset-normalizer == 3.4.4
click == 8.3.1
colorama == 0.4.6
crytography == 46.0.3
firebase-admin == 6.2.0
Flask == 3.0.0
Flask-SQLAlchemy == 3.1.1
gcloud == 0.18.3
google-api-core == 2.28.1
google-api-python-client == 2.187.0
google-auth == 2.43.0
google-cloud-core == 2.5.0
google-cloud-firestore == 2.21.0
google-cloud-storage == 3.6.0
google-crc32c == 1.7.1
google-resumable-media == 2.8.0
googleapis-common-protos == 1.72.0
greenlet == 3.2.4
grpcio == 1.76.0
grpcio-status == 1.62.3
httplib2 == 0.31.0
idna == 3.11
itsdangerous == 2.2.0
Jinja2 == 3.1.3
jwcrypto == 1.5.6
MarkupSafe == 3.0.3
msgpack == 1.1.2
mysql-connector-python == 8.1.0
mysqlclient == 2.2.4
numpy == 1.24.3
oauth2client == 4.1.3
opencv-python == 4.8.1.78
Pillow == 10.0.0
pip == 21.2.3
proto-plus == 1.26.1
protobuf == 4.21.12
pyasn1 == 0.6.1
pyasn1_modules == 0.4.2
pycparser == 2.23
pycryptodome == 3.23.0
PyJWT == 2.10.1
PyMySQL == 1.1.2
pyparsing == 3.2.5
Pyrebase4 == 4.8.0
python-dateutil == 2.8.2
python-dotenv == 1.0.0
python-jwt == 4.1.0
pytz == 2024.1
requests == 2.32.5
requests-toolbelt == 0.10.1
rsa == 4.9.1
setuptools == 57.4.0
six == 1.17.0
SQLAlchemy == 2.0.44
stripe == 7.7.0
typing_extensions == 4.15.0
uritemplate == 4.2.0
urllib3 == 1.26.20
Werkzeug == 3.0.1